[Fix] #136 primary key 와 type 맞춤

+++ backend/src/main/java/com/example/nexus/domain/settlement/SettlementRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface SettlementRepository extends JpaRepository<Settlement, Integer> {
+public interface SettlementRepository extends JpaRepository<Settlement, Long> {
     List<Settlement> findAllByStoreIdx(Long storeIdx);
 }
